backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
pydantic==2.5.3
pydantic-settings==2.1.0
alembic==1.13.1
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
langgraph==0.0.20
openai==1.10.0
httpx==0.26.0
pymongo==4.6.1
motor==3.3.2
faiss-cpu==1.7.4
numpy==1.26.3
pandas==2.1.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
pypdf2==3.0.1
structlog==24.1.0
prometheus-client==0.19.0
